refactor: moved useGlobalTheme to @hooks
