+++ babel.config.js
@@ -14,6 +14,7 @@ module.exports = function (api) {
             '@utils': './src/utils',
             '@theme': './src/theme',
             '@modules': './src/modules',
+            '@hooks': './src/hooks',
           },
         },
       ],
