+++ src/views/PetMedicalRecord/PetMedicalRecord.tsx
@@ -1,5 +1,5 @@
 import { FlatList } from 'react-native-gesture-handler';
-import { useGlobalTheme } from '../../providers/ThemeProvider';
+import { useGlobalTheme } from '@hooks/useGlobalTheme';
 import { View, Text, TouchableOpacity } from 'react-native';
 import { Ionicons } from '@expo/vector-icons';
