+++ src/views/UserProfile/UserProfile.tsx
@@ -1,0 +1,6 @@
+import { View } from 'react-native';
+import { useGlobalTheme } from '@hooks/useGlobalTheme';
+
+export default function UserProfile() {
+  return <View></View>;
+}
