+++ tsconfig.json
@@ -30,6 +30,9 @@
       ],
       "@firebase/auth": [
         "./node_modules/@firebase/auth/dist/index.rn.d.ts"
+      ],
+      "@hooks/*": [
+        "./src/hooks/*"
       ]
     }
   }
